+++ cmd/state_unprotect.go
@@ -32,16 +32,18 @@ import (
 
 func newStateUnprotectCommand() *cobra.Command {
 	var unprotectAll bool
+	var stack string
+
 	cmd := &cobra.Command{
 		Use:   "unprotect",
 		Short: "Unprotect resources in a stack's state",
 		Long: `Unprotect resource in a stack's state
-		
+
 This command clears the 'protect' bit on one or more resources, allowing those resources to be deleted.`,
 		Args: cmdutil.MaximumNArgs(1),
 		Run: cmdutil.RunFunc(func(cmd *cobra.Command, args []string) error {
 			if unprotectAll {
-				return unprotectAllResources()
+				return unprotectAllResources(stack)
 			}
 
 			if len(args) != 1 {
@@ -49,16 +51,19 @@ This command clears the 'protect' bit on one or more resources, allowing those r
 			}
 
 			urn := resource.URN(args[0])
-			return unprotectResource(urn)
+			return unprotectResource(stack, urn)
 		}),
 	}
 
+	cmd.PersistentFlags().StringVarP(
+		&stack, "stack", "s", "",
+		"The name of the stack to operate on. Defaults to the current stack")
 	cmd.Flags().BoolVar(&unprotectAll, "all", false, "Unprotect all resources in the checkpoint")
 	return cmd
 }
 
-func unprotectAllResources() error {
-	err := runTotalStateEdit(func(_ display.Options, snap *deploy.Snapshot) error {
+func unprotectAllResources(stackName string) error {
+	err := runTotalStateEdit(stackName, func(_ display.Options, snap *deploy.Snapshot) error {
 		for _, res := range snap.Resources {
 			err := edit.UnprotectResource(snap, res)
 			contract.AssertNoError(err)
@@ -74,8 +79,8 @@ func unprotectAllResources() error {
 	return nil
 }
 
-func unprotectResource(urn resource.URN) error {
-	err := runStateEdit(urn, edit.UnprotectResource)
+func unprotectResource(stackName string, urn resource.URN) error {
+	err := runStateEdit(stackName, urn, edit.UnprotectResource)
 	if err != nil {
 		return err
 	}