+++ cmd/stack_rm.go
@@ -79,7 +79,7 @@ func newStackRmCmd() *cobra.Command {
 
 			if !preserveConfig {
 				// Blow away stack specific settings if they exist. If we get an ENOENT error, ignore it.
-				if path, err := workspace.DetectProjectStackPath(s.Ref().Name()); err == nil {
+				if path, err := workspace.DetectProjectStackPath(s.Ref().Owner(), s.Ref().Name()); err == nil {
 					if err = os.Remove(path); err != nil && !os.IsNotExist(err) {
 						return err
 					}